% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublets.R
\name{FindBestClusterLinComb}
\alias{FindBestClusterLinComb}
\title{findBestClusterLinComb}
\usage{
FindBestClusterLinComb(clust, counts, clusters, plot = T)
}
\description{
compare all pair of clusters to explain given `clust` by their linear combination.
}
